requirements (1).txt
streamlit
pandas
numpy
scikit-learn
scikit-surprise
